requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
sqlparse==0.5.3
psycopg2-binary
python-decouple
django-filter==25.1
bcrypt
djangorestframework-simplejwt==5.5.1
drf_yasg==1.21.10
